Goodnight commit. Carried out testing to evaluate the feasibility of previous bidirectional relationship commit. It was a success. I intend to use the knowledge gained from the testing to move forward onto implementation now. (I may be needing to carry out similar tests on the OneToMany relationship between Team and Players first however)

+++ src/main/java/Model/DAO/League.java
@@ -36,7 +36,7 @@ public class League
         }
     }
 
-    public static void deleteTeam(int id) // TODO Maybe change to a team object
+    public static void deleteTeam(Long id) // TODO Maybe change to a team object
     {
         EntityManager em = ENTITY_MANAGER_FACTORY.createEntityManager();
         EntityTransaction et = null;
